Move unflatten_tensors function
Which is only used in ConjugateGradientOptimizer.
