+++ src/rl_replicas/utils.py
@@ -46,25 +46,6 @@ def gae(
     return gaes
 
 
-def unflatten_tensors(
-    flattened: np.ndarray, tensor_shapes: List[torch.Size]
-) -> List[Tensor]:
-    """
-    Unflatten a flattened tensors into a list of tensors
-
-    :param flattened: (np.ndarray) Flattened tensors.
-    :param tensor_shapes: (List[torch.Size]) Tensor shapes.
-    :return: (List[np.ndarray]) Unflattened list of tensors.
-    """
-    tensor_sizes = list(map(np.prod, tensor_shapes))
-    indices = np.cumsum(tensor_sizes)[:-1]
-
-    return [
-        np.reshape(pair[0], pair[1])
-        for pair in zip(np.split(flattened, indices), tensor_shapes)
-    ]
-
-
 def polyak_average(
     params: Iterable[nn.Parameter], target_params: Iterable[nn.Parameter], rho: float
 ) -> None:
